Split CompactSortedFiles tests into separate files
