+++ java/compaction-job-execution/src/test/java/sleeper/compaction/jobexecution/CompactSortedFilesIteratorTest.java
@@ -1,0 +1,135 @@
+/*
+ * Copyright 2022 Crown Copyright
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package sleeper.compaction.jobexecution;
+
+import org.junit.Test;
+import sleeper.compaction.job.CompactionJob;
+import sleeper.core.iterator.impl.AgeOffIterator;
+import sleeper.core.partition.PartitionsBuilder;
+import sleeper.core.record.Record;
+import sleeper.core.schema.Schema;
+import sleeper.statestore.StateStore;
+
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.List;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static sleeper.compaction.jobexecution.CompactSortedFilesTestData.readDataFile;
+import static sleeper.compaction.jobexecution.CompactSortedFilesTestData.specifiedFromEvens;
+import static sleeper.compaction.jobexecution.CompactSortedFilesTestData.specifiedFromOdds;
+import static sleeper.compaction.jobexecution.CompactSortedFilesTestUtils.assertReadyForGC;
+import static sleeper.compaction.jobexecution.CompactSortedFilesTestUtils.createCompactSortedFiles;
+
+public class CompactSortedFilesIteratorTest extends CompactSortedFilesTestBase {
+
+    @Test
+    public void filesShouldMergeAndApplyIteratorCorrectlyLongKey() throws Exception {
+        // Given
+        Schema schema = CompactSortedFilesTestUtils.createSchemaWithKeyTimestampValue();
+        StateStore stateStore = createInitStateStore(schema);
+        CompactSortedFilesTestDataHelper dataHelper = new CompactSortedFilesTestDataHelper(schema, stateStore);
+
+        List<Record> data1 = specifiedFromEvens((even, record) -> {
+            record.put("key", (long) even);
+            record.put("timestamp", System.currentTimeMillis());
+            record.put("value", 987654321L);
+        });
+        List<Record> data2 = specifiedFromOdds((odd, record) -> {
+            record.put("key", (long) odd);
+            record.put("timestamp", 0L);
+            record.put("value", 123456789L);
+        });
+        dataHelper.writeLeafFile(folderName + "/file1.parquet", data1, 0L, 198L);
+        dataHelper.writeLeafFile(folderName + "/file2.parquet", data2, 1L, 199L);
+
+        CompactionJob compactionJob = compactionFactoryBuilder()
+                .iteratorClassName(AgeOffIterator.class.getName())
+                .iteratorConfig("timestamp,1000000")
+                .build().createCompactionJob(dataHelper.allFileInfos(), dataHelper.singlePartition().getId());
+        dataHelper.addFilesToStateStoreForJob(compactionJob);
+
+        // When
+        CompactSortedFiles compactSortedFiles = createCompactSortedFiles(schema, compactionJob, stateStore);
+        CompactSortedFiles.CompactionJobSummary summary = compactSortedFiles.compact();
+
+        // Then
+        //  - Read output files and check that they contain the right results
+        assertThat(summary.getLinesRead()).isEqualTo(200L);
+        assertThat(summary.getLinesWritten()).isEqualTo(100L);
+        assertThat(readDataFile(schema, compactionJob.getOutputFile())).isEqualTo(data1);
+
+        // - Check DynamoDBStateStore has correct ready for GC files
+        assertReadyForGC(stateStore, dataHelper.allFileInfos());
+
+        // - Check DynamoDBStateStore has correct active files
+        assertThat(stateStore.getActiveFiles())
+                .usingRecursiveFieldByFieldElementComparatorIgnoringFields("lastStateStoreUpdateTime")
+                .containsExactly(dataHelper.expectedLeafFile(compactionJob.getOutputFile(), 100L, 0L, 198L));
+    }
+
+    @Test
+    public void filesShouldMergeAndSplitAndApplyIteratorCorrectlyLongKey() throws Exception {
+        // Given
+        Schema schema = CompactSortedFilesTestUtils.createSchemaWithKeyTimestampValue();
+        StateStore stateStore = createStateStore(schema);
+        stateStore.initialise(new PartitionsBuilder(schema)
+                .leavesWithSplits(Arrays.asList("A", "B"), Collections.singletonList(100L))
+                .parentJoining("C", "A", "B")
+                .buildList());
+        CompactSortedFilesTestDataHelper dataHelper = new CompactSortedFilesTestDataHelper(schema, stateStore);
+
+        List<Record> data1 = specifiedFromEvens((even, record) -> {
+            record.put("key", (long) even);
+            record.put("timestamp", System.currentTimeMillis());
+            record.put("value", 987654321L);
+        });
+        List<Record> data2 = specifiedFromOdds((odd, record) -> {
+            record.put("key", (long) odd);
+            record.put("timestamp", 0L);
+            record.put("value", 123456789L);
+        });
+        dataHelper.writeRootFile(folderName + "/file1.parquet", data1, 0L, 198L);
+        dataHelper.writeRootFile(folderName + "/file2.parquet", data2, 1L, 199L);
+
+        CompactionJob compactionJob = compactionFactoryBuilder()
+                .iteratorClassName(AgeOffIterator.class.getName())
+                .iteratorConfig("timestamp,1000000")
+                .build().createSplittingCompactionJob(dataHelper.allFileInfos(), "C", "A", "B", 100L, 0);
+        dataHelper.addFilesToStateStoreForJob(compactionJob);
+
+        // When
+        CompactSortedFiles compactSortedFiles = createCompactSortedFiles(schema, compactionJob, stateStore);
+        CompactSortedFiles.CompactionJobSummary summary = compactSortedFiles.compact();
+
+        // Then
+        //  - Read output files and check that they contain the right results
+        assertThat(summary.getLinesRead()).isEqualTo(200L);
+        assertThat(summary.getLinesWritten()).isEqualTo(100L);
+        assertThat(readDataFile(schema, compactionJob.getOutputFiles().getLeft())).isEqualTo(data1.subList(0, 50));
+        assertThat(readDataFile(schema, compactionJob.getOutputFiles().getRight())).isEqualTo(data1.subList(50, 100));
+
+        // - Check DynamoDBStateStore has correct ready for GC files
+        assertReadyForGC(stateStore, dataHelper.allFileInfos());
+
+        // - Check DynamoDBStateStore has correct active files
+        assertThat(stateStore.getActiveFiles())
+                .usingRecursiveFieldByFieldElementComparatorIgnoringFields("lastStateStoreUpdateTime")
+                .containsExactlyInAnyOrder(
+                        dataHelper.expectedPartitionFile("A", compactionJob.getOutputFiles().getLeft(), 50L, 0L, 98L),
+                        dataHelper.expectedPartitionFile("B", compactionJob.getOutputFiles().getRight(), 50L, 100L, 198L));
+    }
+}
