+++ java/compaction-job-execution/src/test/java/sleeper/compaction/jobexecution/CompactSortedFilesTest.java
@@ -17,18 +17,13 @@ package sleeper.compaction.jobexecution;
 
 import org.junit.Test;
 import sleeper.compaction.job.CompactionJob;
-import sleeper.core.iterator.impl.AgeOffIterator;
-import sleeper.core.partition.PartitionsBuilder;
 import sleeper.core.record.Record;
-import sleeper.core.schema.Field;
 import sleeper.core.schema.Schema;
 import sleeper.core.schema.type.ByteArrayType;
 import sleeper.core.schema.type.LongType;
 import sleeper.core.schema.type.StringType;
 import sleeper.statestore.StateStore;
 
-import java.util.Arrays;
-import java.util.Collections;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -41,13 +36,8 @@ import static sleeper.compaction.jobexecution.CompactSortedFilesTestData.keyAndT
 import static sleeper.compaction.jobexecution.CompactSortedFilesTestData.keyAndTwoValuesSortedOddLongs;
 import static sleeper.compaction.jobexecution.CompactSortedFilesTestData.keyAndTwoValuesSortedOddStrings;
 import static sleeper.compaction.jobexecution.CompactSortedFilesTestData.readDataFile;
-import static sleeper.compaction.jobexecution.CompactSortedFilesTestData.specifiedAndTwoValuesFromEvens;
-import static sleeper.compaction.jobexecution.CompactSortedFilesTestData.specifiedAndTwoValuesFromOdds;
-import static sleeper.compaction.jobexecution.CompactSortedFilesTestData.specifiedFromEvens;
-import static sleeper.compaction.jobexecution.CompactSortedFilesTestData.specifiedFromOdds;
 import static sleeper.compaction.jobexecution.CompactSortedFilesTestUtils.assertReadyForGC;
 import static sleeper.compaction.jobexecution.CompactSortedFilesTestUtils.createCompactSortedFiles;
-import static sleeper.compaction.jobexecution.CompactSortedFilesTestUtils.createSchemaWithTwoTypedValuesAndKeyFields;
 import static sleeper.compaction.jobexecution.CompactSortedFilesTestUtils.createSchemaWithTypesForKeyAndTwoValues;
 
 public class CompactSortedFilesTest extends CompactSortedFilesTestBase {
@@ -197,356 +187,5 @@ public class CompactSortedFilesTest extends CompactSortedFilesTestBase {
         assertThat(stateStore.getActiveFiles())
                 .usingRecursiveFieldByFieldElementComparatorIgnoringFields("lastStateStoreUpdateTime")
                 .containsExactly(dataHelper.expectedLeafFile(compactionJob.getOutputFile(), 200L, new byte[]{0, 0}, new byte[]{1, 71}));
-    }
-
-    @Test
-    public void filesShouldMergeCorrectlyWhenSomeAreEmpty() throws Exception {
-        // Given
-        Schema schema = createSchemaWithTypesForKeyAndTwoValues(new LongType(), new LongType(), new LongType());
-        StateStore stateStore = createInitStateStore(schema);
-        CompactSortedFilesTestDataHelper dataHelper = new CompactSortedFilesTestDataHelper(schema, stateStore);
-
-        List<Record> data = keyAndTwoValuesSortedEvenLongs();
-        dataHelper.writeLeafFile(folderName + "/file1.parquet", data, 0L, 198L);
-        dataHelper.writeLeafFile(folderName + "/file2.parquet", Collections.emptyList(), null, null);
-
-        CompactionJob compactionJob = compactionFactory().createCompactionJob(
-                dataHelper.allFileInfos(), dataHelper.singlePartition().getId());
-        dataHelper.addFilesToStateStoreForJob(compactionJob);
-
-        // When
-        CompactSortedFiles compactSortedFiles = createCompactSortedFiles(schema, compactionJob, stateStore);
-        CompactSortedFiles.CompactionJobSummary summary = compactSortedFiles.compact();
-
-        // Then
-        //  - Read output file and check that it contains the right results
-        assertThat(summary.getLinesRead()).isEqualTo(data.size());
-        assertThat(summary.getLinesWritten()).isEqualTo(data.size());
-        assertThat(readDataFile(schema, compactionJob.getOutputFile())).isEqualTo(data);
-
-        // - Check DynamoDBStateStore has correct ready for GC files
-        assertReadyForGC(stateStore, dataHelper.allFileInfos());
-
-        // - Check DynamoDBStateStore has correct active files
-        assertThat(stateStore.getActiveFiles())
-                .usingRecursiveFieldByFieldElementComparatorIgnoringFields("lastStateStoreUpdateTime")
-                .containsExactly(dataHelper.expectedLeafFile(compactionJob.getOutputFile(), 100L, 0L, 198L));
-    }
-
-    @Test
-    public void filesShouldMergeCorrectlyWhenAllAreEmpty() throws Exception {
-        // Given
-        Schema schema = createSchemaWithTypesForKeyAndTwoValues(new LongType(), new LongType(), new LongType());
-        StateStore stateStore = createInitStateStore(schema);
-        CompactSortedFilesTestDataHelper dataHelper = new CompactSortedFilesTestDataHelper(schema, stateStore);
-
-        dataHelper.writeLeafFile(folderName + "/file1.parquet", Collections.emptyList(), null, null);
-        dataHelper.writeLeafFile(folderName + "/file2.parquet", Collections.emptyList(), null, null);
-
-        CompactionJob compactionJob = compactionFactory().createCompactionJob(
-                dataHelper.allFileInfos(), dataHelper.singlePartition().getId());
-        dataHelper.addFilesToStateStoreForJob(compactionJob);
-
-        // When
-        CompactSortedFiles compactSortedFiles = createCompactSortedFiles(schema, compactionJob, stateStore);
-        CompactSortedFiles.CompactionJobSummary summary = compactSortedFiles.compact();
-
-        // Then
-        //  - Read output file and check that it contains the right results
-        assertThat(summary.getLinesRead()).isZero();
-        assertThat(summary.getLinesWritten()).isZero();
-        assertThat(readDataFile(schema, compactionJob.getOutputFile())).isEmpty();
-
-        // - Check DynamoDBStateStore has correct ready for GC files
-        assertReadyForGC(stateStore, dataHelper.allFileInfos());
-
-        // - Check DynamoDBStateStore has correct active files
-        assertThat(stateStore.getActiveFiles())
-                .usingRecursiveFieldByFieldElementComparatorIgnoringFields("lastStateStoreUpdateTime")
-                .containsExactly(dataHelper.expectedLeafFile(compactionJob.getOutputFile(), 0L, null, null));
-    }
-
-    @Test
-    public void filesShouldMergeAndSplitCorrectlyAndDynamoUpdated() throws Exception {
-        // Given
-        Schema schema = createSchemaWithTypesForKeyAndTwoValues(new LongType(), new LongType(), new LongType());
-        StateStore stateStore = createStateStore(schema);
-        stateStore.initialise(new PartitionsBuilder(schema)
-                .leavesWithSplits(Arrays.asList("A", "B"), Collections.singletonList(100L))
-                .parentJoining("C", "A", "B")
-                .buildList());
-        CompactSortedFilesTestDataHelper dataHelper = new CompactSortedFilesTestDataHelper(schema, stateStore);
-
-        List<Record> data1 = keyAndTwoValuesSortedEvenLongs();
-        List<Record> data2 = keyAndTwoValuesSortedOddLongs();
-        dataHelper.writeRootFile(folderName + "/file1.parquet", data1, 0L, 198L);
-        dataHelper.writeRootFile(folderName + "/file2.parquet", data2, 1L, 199L);
-
-        CompactionJob compactionJob = compactionFactory().createSplittingCompactionJob(
-                dataHelper.allFileInfos(), "C", "A", "B", 100L, 0);
-        dataHelper.addFilesToStateStoreForJob(compactionJob);
-
-        // When
-        CompactSortedFiles compactSortedFiles = createCompactSortedFiles(schema, compactionJob, stateStore);
-        CompactSortedFiles.CompactionJobSummary summary = compactSortedFiles.compact();
-
-        // Then
-        //  - Read output files and check that they contain the right results
-        List<Record> expectedResults = combineSortedBySingleKey(data1, data2);
-        assertThat(summary.getLinesRead()).isEqualTo(200L);
-        assertThat(summary.getLinesWritten()).isEqualTo(200L);
-        assertThat(readDataFile(schema, compactionJob.getOutputFiles().getLeft())).isEqualTo(expectedResults.subList(0, 100));
-        assertThat(readDataFile(schema, compactionJob.getOutputFiles().getRight())).isEqualTo(expectedResults.subList(100, 200));
-
-        // - Check DynamoDBStateStore has correct ready for GC files
-        assertReadyForGC(stateStore, dataHelper.allFileInfos());
-
-        // - Check DynamoDBStateStore has correct active files
-        assertThat(stateStore.getActiveFiles())
-                .usingRecursiveFieldByFieldElementComparatorIgnoringFields("lastStateStoreUpdateTime")
-                .containsExactlyInAnyOrder(
-                        dataHelper.expectedPartitionFile("A", compactionJob.getOutputFiles().getLeft(), 100L, 0L, 99L),
-                        dataHelper.expectedPartitionFile("B", compactionJob.getOutputFiles().getRight(), 100L, 100L, 199L));
-    }
-
-    @Test
-    public void filesShouldMergeAndSplitCorrectlyWith2DimKeySplitOnFirstKey() throws Exception {
-        // Given
-        Field field1 = new Field("key1", new LongType());
-        Field field2 = new Field("key2", new StringType());
-        Schema schema = createSchemaWithTwoTypedValuesAndKeyFields(new LongType(), new LongType(), field1, field2);
-        StateStore stateStore = createStateStore(schema);
-        stateStore.initialise(new PartitionsBuilder(schema)
-                .leavesWithSplits(Arrays.asList("A", "B"), Collections.singletonList(100L))
-                .parentJoining("C", "A", "B")
-                .buildList());
-        CompactSortedFilesTestDataHelper dataHelper = new CompactSortedFilesTestDataHelper(schema, stateStore);
-
-        List<Record> data1 = specifiedAndTwoValuesFromEvens((even, record) -> {
-            record.put(field1.getName(), (long) even);
-            record.put(field2.getName(), "A");
-        });
-        List<Record> data2 = specifiedAndTwoValuesFromOdds((odd, record) -> {
-            record.put(field1.getName(), (long) odd);
-            record.put(field2.getName(), "A");
-        });
-        dataHelper.writeRootFile(folderName + "/file1.parquet", data1, 0L, 198L);
-        dataHelper.writeRootFile(folderName + "/file2.parquet", data2, 1L, 199L);
-
-        CompactionJob compactionJob = compactionFactory().createSplittingCompactionJob(
-                dataHelper.allFileInfos(), "C", "A", "B", 100L, 0);
-        dataHelper.addFilesToStateStoreForJob(compactionJob);
-
-        // When
-        CompactSortedFiles compactSortedFiles = createCompactSortedFiles(schema, compactionJob, stateStore);
-        CompactSortedFiles.CompactionJobSummary summary = compactSortedFiles.compact();
-
-        // Then
-        //  - Read output files and check that they contain the right results
-        List<Record> expectedResults = combineSortedBySingleKey(data1, data2, record -> record.get(field1.getName()));
-        assertThat(summary.getLinesRead()).isEqualTo(200L);
-        assertThat(summary.getLinesWritten()).isEqualTo(200L);
-        assertThat(readDataFile(schema, compactionJob.getOutputFiles().getLeft())).isEqualTo(expectedResults.subList(0, 100));
-        assertThat(readDataFile(schema, compactionJob.getOutputFiles().getRight())).isEqualTo(expectedResults.subList(100, 200));
-
-        // - Check DynamoDBStateStore has correct ready for GC files
-        assertReadyForGC(stateStore, dataHelper.allFileInfos());
-
-        // - Check DynamoDBStateStore has correct active files
-        assertThat(stateStore.getActiveFiles())
-                .usingRecursiveFieldByFieldElementComparatorIgnoringFields("lastStateStoreUpdateTime")
-                .containsExactlyInAnyOrder(
-                        dataHelper.expectedPartitionFile("A", compactionJob.getOutputFiles().getLeft(), 100L, 0L, 99L),
-                        dataHelper.expectedPartitionFile("B", compactionJob.getOutputFiles().getRight(), 100L, 100L, 199L));
-    }
-
-    @Test
-    public void filesShouldMergeAndSplitCorrectlyWith2DimKeySplitOnSecondKey() throws Exception {
-        // Given
-        Field field1 = new Field("key1", new LongType());
-        Field field2 = new Field("key2", new StringType());
-        Schema schema = createSchemaWithTwoTypedValuesAndKeyFields(new LongType(), new LongType(), field1, field2);
-        StateStore stateStore = createStateStore(schema);
-        stateStore.initialise(new PartitionsBuilder(schema)
-                .leavesWithSplitsOnDimension(1, Arrays.asList("A", "B"), Collections.singletonList("A2"))
-                .parentJoining("C", "A", "B")
-                .buildList());
-        CompactSortedFilesTestDataHelper dataHelper = new CompactSortedFilesTestDataHelper(schema, stateStore);
-
-        List<Record> data1 = specifiedAndTwoValuesFromEvens((even, record) -> {
-            record.put(field1.getName(), (long) even);
-            record.put(field2.getName(), "A");
-        });
-        List<Record> data2 = specifiedAndTwoValuesFromOdds((odd, record) -> {
-            record.put(field1.getName(), (long) odd);
-            record.put(field2.getName(), "B");
-        });
-        dataHelper.writeRootFile(folderName + "/file1.parquet", data1, 0L, 198L);
-        dataHelper.writeRootFile(folderName + "/file2.parquet", data2, 1L, 199L);
-
-        CompactionJob compactionJob = compactionFactory().createSplittingCompactionJob(
-                dataHelper.allFileInfos(), "C", "A", "B", "A2", 1);
-        dataHelper.addFilesToStateStoreForJob(compactionJob);
-
-        // When
-        CompactSortedFiles compactSortedFiles = createCompactSortedFiles(schema, compactionJob, stateStore);
-        CompactSortedFiles.CompactionJobSummary summary = compactSortedFiles.compact();
-
-        // Then
-        //  - Read output files and check that they contain the right results
-        assertThat(summary.getLinesRead()).isEqualTo(200L);
-        assertThat(summary.getLinesWritten()).isEqualTo(200L);
-        assertThat(readDataFile(schema, compactionJob.getOutputFiles().getLeft())).isEqualTo(data1);
-        assertThat(readDataFile(schema, compactionJob.getOutputFiles().getRight())).isEqualTo(data2);
-
-        // - Check DynamoDBStateStore has correct ready for GC files
-        assertReadyForGC(stateStore, dataHelper.allFileInfos());
-
-        // - Check DynamoDBStateStore has correct active files
-        assertThat(stateStore.getActiveFiles())
-                .usingRecursiveFieldByFieldElementComparatorIgnoringFields("lastStateStoreUpdateTime")
-                .containsExactlyInAnyOrder(
-                        dataHelper.expectedPartitionFile("A", compactionJob.getOutputFiles().getLeft(), 100L, 0L, 198L),
-                        dataHelper.expectedPartitionFile("B", compactionJob.getOutputFiles().getRight(), 100L, 1L, 199L));
-    }
-
-    @Test
-    public void filesShouldMergeAndSplitCorrectlyWhenOneChildFileIsEmpty() throws Exception {
-        // Given
-        Schema schema = createSchemaWithTypesForKeyAndTwoValues(new LongType(), new LongType(), new LongType());
-        StateStore stateStore = createStateStore(schema);
-        stateStore.initialise(new PartitionsBuilder(schema)
-                .leavesWithSplits(Arrays.asList("A", "B"), Collections.singletonList(200L))
-                .parentJoining("C", "A", "B")
-                .buildList());
-        CompactSortedFilesTestDataHelper dataHelper = new CompactSortedFilesTestDataHelper(schema, stateStore);
-
-        List<Record> data1 = keyAndTwoValuesSortedEvenLongs();
-        List<Record> data2 = keyAndTwoValuesSortedOddLongs();
-        dataHelper.writeRootFile(folderName + "/file1.parquet", data1, 0L, 198L);
-        dataHelper.writeRootFile(folderName + "/file2.parquet", data2, 1L, 199L);
-
-        CompactionJob compactionJob = compactionFactory().createSplittingCompactionJob(
-                dataHelper.allFileInfos(), "C", "A", "B", 200L, 0);
-        dataHelper.addFilesToStateStoreForJob(compactionJob);
-
-        // When
-        CompactSortedFiles compactSortedFiles = createCompactSortedFiles(schema, compactionJob, stateStore);
-        CompactSortedFiles.CompactionJobSummary summary = compactSortedFiles.compact();
-
-        // Then
-        //  - Read output files and check that they contain the right results
-        List<Record> expectedResults = combineSortedBySingleKey(data1, data2);
-        assertThat(summary.getLinesRead()).isEqualTo(200L);
-        assertThat(summary.getLinesWritten()).isEqualTo(200L);
-        assertThat(readDataFile(schema, compactionJob.getOutputFiles().getLeft())).isEqualTo(expectedResults);
-        assertThat(readDataFile(schema, compactionJob.getOutputFiles().getRight())).isEmpty();
-
-        // - Check DynamoDBStateStore has correct ready for GC files
-        assertReadyForGC(stateStore, dataHelper.allFileInfos());
-
-        // - Check DynamoDBStateStore has correct active files
-        assertThat(stateStore.getActiveFiles())
-                .usingRecursiveFieldByFieldElementComparatorIgnoringFields("lastStateStoreUpdateTime")
-                .containsExactlyInAnyOrder(
-                        dataHelper.expectedPartitionFile("A", compactionJob.getOutputFiles().getLeft(), 200L, 0L, 199L),
-                        dataHelper.expectedPartitionFile("B", compactionJob.getOutputFiles().getRight(), 0L, null, null));
-    }
-
-    @Test
-    public void filesShouldMergeAndApplyIteratorCorrectlyLongKey() throws Exception {
-        // Given
-        Schema schema = CompactSortedFilesTestUtils.createSchemaWithKeyTimestampValue();
-        StateStore stateStore = createInitStateStore(schema);
-        CompactSortedFilesTestDataHelper dataHelper = new CompactSortedFilesTestDataHelper(schema, stateStore);
-
-        List<Record> data1 = specifiedFromEvens((even, record) -> {
-            record.put("key", (long) even);
-            record.put("timestamp", System.currentTimeMillis());
-            record.put("value", 987654321L);
-        });
-        List<Record> data2 = specifiedFromOdds((odd, record) -> {
-            record.put("key", (long) odd);
-            record.put("timestamp", 0L);
-            record.put("value", 123456789L);
-        });
-        dataHelper.writeLeafFile(folderName + "/file1.parquet", data1, 0L, 198L);
-        dataHelper.writeLeafFile(folderName + "/file2.parquet", data2, 1L, 199L);
-
-        CompactionJob compactionJob = compactionFactoryBuilder()
-                .iteratorClassName(AgeOffIterator.class.getName())
-                .iteratorConfig("timestamp,1000000")
-                .build().createCompactionJob(dataHelper.allFileInfos(), dataHelper.singlePartition().getId());
-        dataHelper.addFilesToStateStoreForJob(compactionJob);
-
-        // When
-        CompactSortedFiles compactSortedFiles = createCompactSortedFiles(schema, compactionJob, stateStore);
-        CompactSortedFiles.CompactionJobSummary summary = compactSortedFiles.compact();
-
-        // Then
-        //  - Read output files and check that they contain the right results
-        assertThat(summary.getLinesRead()).isEqualTo(200L);
-        assertThat(summary.getLinesWritten()).isEqualTo(100L);
-        assertThat(readDataFile(schema, compactionJob.getOutputFile())).isEqualTo(data1);
-
-        // - Check DynamoDBStateStore has correct ready for GC files
-        assertReadyForGC(stateStore, dataHelper.allFileInfos());
-
-        // - Check DynamoDBStateStore has correct active files
-        assertThat(stateStore.getActiveFiles())
-                .usingRecursiveFieldByFieldElementComparatorIgnoringFields("lastStateStoreUpdateTime")
-                .containsExactly(dataHelper.expectedLeafFile(compactionJob.getOutputFile(), 100L, 0L, 198L));
-    }
-
-    @Test
-    public void filesShouldMergeAndSplitAndApplyIteratorCorrectlyLongKey() throws Exception {
-        // Given
-        Schema schema = CompactSortedFilesTestUtils.createSchemaWithKeyTimestampValue();
-        StateStore stateStore = createStateStore(schema);
-        stateStore.initialise(new PartitionsBuilder(schema)
-                .leavesWithSplits(Arrays.asList("A", "B"), Collections.singletonList(100L))
-                .parentJoining("C", "A", "B")
-                .buildList());
-        CompactSortedFilesTestDataHelper dataHelper = new CompactSortedFilesTestDataHelper(schema, stateStore);
-
-        List<Record> data1 = specifiedFromEvens((even, record) -> {
-            record.put("key", (long) even);
-            record.put("timestamp", System.currentTimeMillis());
-            record.put("value", 987654321L);
-        });
-        List<Record> data2 = specifiedFromOdds((odd, record) -> {
-            record.put("key", (long) odd);
-            record.put("timestamp", 0L);
-            record.put("value", 123456789L);
-        });
-        dataHelper.writeRootFile(folderName + "/file1.parquet", data1, 0L, 198L);
-        dataHelper.writeRootFile(folderName + "/file2.parquet", data2, 1L, 199L);
-
-        CompactionJob compactionJob = compactionFactoryBuilder()
-                .iteratorClassName(AgeOffIterator.class.getName())
-                .iteratorConfig("timestamp,1000000")
-                .build().createSplittingCompactionJob(dataHelper.allFileInfos(), "C", "A", "B", 100L, 0);
-        dataHelper.addFilesToStateStoreForJob(compactionJob);
-
-        // When
-        CompactSortedFiles compactSortedFiles = createCompactSortedFiles(schema, compactionJob, stateStore);
-        CompactSortedFiles.CompactionJobSummary summary = compactSortedFiles.compact();
-
-        // Then
-        //  - Read output files and check that they contain the right results
-        assertThat(summary.getLinesRead()).isEqualTo(200L);
-        assertThat(summary.getLinesWritten()).isEqualTo(100L);
-        assertThat(readDataFile(schema, compactionJob.getOutputFiles().getLeft())).isEqualTo(data1.subList(0, 50));
-        assertThat(readDataFile(schema, compactionJob.getOutputFiles().getRight())).isEqualTo(data1.subList(50, 100));
-
-        // - Check DynamoDBStateStore has correct ready for GC files
-        assertReadyForGC(stateStore, dataHelper.allFileInfos());
-
-        // - Check DynamoDBStateStore has correct active files
-        assertThat(stateStore.getActiveFiles())
-                .usingRecursiveFieldByFieldElementComparatorIgnoringFields("lastStateStoreUpdateTime")
-                .containsExactlyInAnyOrder(
-                        dataHelper.expectedPartitionFile("A", compactionJob.getOutputFiles().getLeft(), 50L, 0L, 98L),
-                        dataHelper.expectedPartitionFile("B", compactionJob.getOutputFiles().getRight(), 50L, 100L, 198L));
     }
 }
